Rename ast files to compiler, will create a new ast file when switching to variants
