+++ src/anzu.m.cpp
@@ -1,9 +1,9 @@
 #include "stack_frame.hpp"
 #include "op_codes.hpp"
+#include "print.hpp"
 #include "lexer.hpp"
 #include "parser.hpp"
-#include "print.hpp"
-#include "ast.hpp"
+#include "compiler.hpp"
 
 #include <string>
 #include <variant>
