+++ src/compiler.cpp
@@ -1,4 +1,4 @@
-#include "ast.hpp"
+#include "compiler.hpp"
 #include "lexer.hpp"
 #include "object.hpp"
 #include "print.hpp"
